backend/requirements.txt
Flask==2.3.3
flask_cors==4.0.0
mysql-connector==2.2.9
